[#11290] log4j2-it, extract PatternTestBase
